Struct concordium_wasm::artifact::Artifact

pub struct Artifact<ImportFunc, CompiledCode> {
    pub imports: Vec<ImportFunc>,
    pub ty: Vec<FunctionType>,
    pub table: InstantiatedTable,
    pub memory: Option<ArtifactMemory>,
    pub global: InstantiatedGlobals,
    pub export: BTreeMap<Name, FuncIndex>,
    pub code: Vec<CompiledCode>,
}

A processed Wasm module. This no longer has custom sections since they are not needed for further processing. The type parameter ImportFunc is instantiated with the representation of host functions. To efficiently and relatively safely execute the module we preprocess imported functions into an enum. However for testing we sometimes just use raw imports. This type parameter allows us flexibility.

The type parameter CompiledCode is used to allow flexibility in code representation. For testing uses it is convenient that the type is “owned”, in the sense of it being a vector of instructions. For efficient execution, and to avoid deserialization, the code is represented as a byte array (i.e., as a slice of bytes &[u8]) when we execute it after looking the code up from the database.

Fields
imports: Vec<ImportFunc>

Imports by (module name, item name).

ty: Vec<FunctionType>

Types of the module. These are needed for dynamic dispatch, i.e., call-indirect.

table: InstantiatedTable

A fully instantiated table.

memory: Option<ArtifactMemory>

The memory of the artifact.

global: InstantiatedGlobals

Globals initialized with initial values.

export: BTreeMap<Name, FuncIndex>

The exported functions. Validation should ensure that an exported function is a defined one, and not one of the imported ones. Thus the index refers to the index in the code section.

code: Vec<CompiledCode>

The list of functions in the module.

Implementations

impl<I: TryFromImport, R: RunnableCode> Artifact<I, R>

pub fn run<Q: Display + Ord + ?Sized, H: Host<I>>(
&self,
host: &mut H,
name: &Q,
args: &[Value]
) -> RunResult<ExecutionOutcome<H::Interrupt>> where
Name: Borrow<Q>,

Attempt to run the entrypoint using the supplied arguments. The arguments are

host - the structure that resolves calls to external, host, functions.

name - the name of the entrypoint to invoke

args - a list of arguments to the entrypoint. The argument list must match the declared type of the entrypoint. If this is not the case then execution will fail with an error.

Note that this method at present cannot be used to directly call a host function. Only an entrypoint defined in the Wasm module can be called.

pub fn has_entrypoint<Q>(&self, name: &Q) -> bool where
Q: Display + Ord + ?Sized,
Name: Borrow<Q>,

Returns if the given entrypoint name exists in the artifact.

pub fn run_config<H: Host<I>>(
&self,
host: &mut H,
config: RunConfig
) -> RunResult<ExecutionOutcome<H::Interrupt>>

Run a configuration using the provided host to resolve calls to external functions.

This executes code either until

the execution terminates with a result

there is an error

the host triggers an interrupt as a result of an external call.