Trait concordium_wasm::artifact::RunnableCode

pub trait RunnableCode {
    fn num_params(&self) -> u32;

    fn type_idx(&self) -> TypeIndex;

    fn return_type(&self) -> BlockType;

    fn params(&self) -> &[ValueType];

    fn num_locals(&self) -> u32;

    fn locals(&self) -> LocalsIterator<'_>;

    fn code(&self) -> &[u8];
}

A trait encapsulating the properties that are needed to run a function. This trait exists because we have two different kinds of code we run. A fully deserialized code, i.e., where instructions are essentially Vec<InternalOpcode> or we execute directly from &[u8] if the origin of the code is a serialized structure, such as an Artifact retrieved from a database.

Required methods

fn num_params(&self) -> u32

The number of parameters of the function.

fn type_idx(&self) -> TypeIndex

The type of the function, as an index into the list of types of the module.

fn return_type(&self) -> BlockType

The return type of the function.

fn params(&self) -> &[ValueType]

The types of function parameters.

fn num_locals(&self) -> u32

The number of locals declared by the function. This does not include the function parameters, only declared locals.

fn locals(&self) -> LocalsIterator<'_>

An iterator over the locals (not including function parameters).

fn code(&self) -> &[u8]

A reference to the instructions to execute.