Module concordium_wasm::machine

An implementation of the abstract machine that can run artifacts. This module defines types related to code execution. The functions to run code are defined as methods on the Artifact type, e.g., Artifact::run.

Structs

RunConfig

Configuration that can be run. This maintains the snapshot of the state of the machine, such as the current instruction pointer, current memory contents, the function stack, etc.

RuntimeStack

A runtime stack. This contains both the stack in a function, as well as all the function parameters and locals of the function.

Enums

ExecutionOutcome

A successful outcome of code execution.

NoInterrupt

An empty type used when no interrupt is possible by a host function call.

RuntimeError

A runtime error that we impose on top of the Wasm spec.

Value

A Wasm typed value. The values are not inherently signed or unsigned, but we choose signed integers as the representation type.

Traits

Host

The host that can process calls to external, host, functions. This is a Wasm concept. Wasm modules are self-contained and instructions can only modify the Wasm memory and stack and cannot access information about the external world, such as current time for example. Host functions fill that role.

Type Definitions

RunResult

Result of execution. Runtime exceptions are returned as Err(_). This includes traps, illegal memory accesses, etc.