Function concordium_wasm::output::write_custom_section

pub fn write_custom_section(
    out: &mut impl Write, 
    cs: &CustomSection<'_>
) -> OutResult<()>

Output a custom section into the given writer. If the writer already contains a valid Wasm module then a custom section can always be appended, and it will result in another valid Wasm module.