Trait concordium_wasm::parse::Parseable

pub trait Parseable<'a, Ctx>: Sized {
    fn parse(ctx: Ctx, cursor: &mut Cursor<&'a [u8]>) -> ParseResult<Self>;
}

A trait for parsing data. The lifetime and context are useful when we want to parse data without copying, which is useful to avoid copying all the unparsed sections.

The implementors of this trait will generally parse data according to the format specified by Wasm specification.

Required methods

fn parse(ctx: Ctx, cursor: &mut Cursor<&'a [u8]>) -> ParseResult<Self>

Read a value from the cursor, or signal error. This function is responsible for advancing the cursor in-line with the data it has read.