Function concordium_wasm::validate::validate

pub fn validate<O: Borrow<OpCode>, H: Handler<O>>(
    context: &impl HasValidationContext, 
    opcodes: impl Iterator<Item = ParseResult<O>>, 
    handler: H
) -> ValidateResult<H::Outcome>

Validate a single function. In order that this function is as flexible as possible it takes as input just an iterator over opcodes. The function will terminate at the first opcode it fails to read. Validation will ensure that the iterator is fully consumed and properly terminated by an End opcode. The return value is the outcome determined by the handler, as well as the maximum reachable stack height in this function.