Trait concordium_wasm::validate::Handler

pub trait Handler<O> {
    type Outcome: Sized;
    fn handle_opcode(
        &mut self, 
        state: &ValidationState, 
        stack_heigh: usize, 
        opcode: O
    ) -> Result<()>;

    fn finish(self, state: &ValidationState) -> Result<Self::Outcome>;
}

Trait to handle the results of validation. The type parameter should be instantiated with an opcode. The reason it is a type parameter is to support both opcodes and references to opcodes as parameters. The latter is useful because opcodes are not copyable.

Associated Types

type Outcome: Sized

Required methods

fn handle_opcode(
&mut self,
state: &ValidationState,
stack_heigh: usize,
opcode: O
) -> Result<()>

Handle the opcode. This function is called after the validate function itself processes the opcode. Hence the validation state is already updated. However the function does get access to the stack height before the opcode is processed.

fn finish(self, state: &ValidationState) -> Result<Self::Outcome>

Finish processing the code. This function is called after the code body has been successfully validated.